Use recommended go repository structure
